Rename Cypher query files to be more human-readable
