+++ src/neo4j/cypher-queries/company/show/index.js
@@ -1,19 +1,19 @@
 import getShowQuery from './show';
 import getShowAwardsQuery from './show-awards';
-import getShowAwardsSubsequentVersionMaterialQuery from './show-awards-subsequent-version-material';
-import getShowAwardsSourcingMaterialQuery from './show-awards-sourcing-material';
-import getShowAwardsRightsGrantorMaterialQuery from './show-awards-rights-grantor-material';
-import getShowMaterialProductionsQuery from './show-material-productions';
+import getShowAwardsForSubsequentVersionMaterialsQuery from './show-awards-for-subsequent-version-materials';
+import getShowAwardsForSourcingMaterialsQuery from './show-awards-for-sourcing-materials';
+import getShowAwardsForRightsGrantorMaterialsQuery from './show-awards-for-rights-grantor-materials';
 import getShowMaterialsQuery from './show-materials';
+import getShowProductionsOfMaterialsQuery from './show-productions-of-materials';
 import getShowProductionsQuery from './show-productions';
 
 export default () => [
 	getShowQuery(),
 	getShowAwardsQuery(),
-	getShowAwardsSubsequentVersionMaterialQuery(),
-	getShowAwardsSourcingMaterialQuery(),
-	getShowAwardsRightsGrantorMaterialQuery(),
-	getShowMaterialProductionsQuery(),
+	getShowAwardsForSubsequentVersionMaterialsQuery(),
+	getShowAwardsForSourcingMaterialsQuery(),
+	getShowAwardsForRightsGrantorMaterialsQuery(),
 	getShowMaterialsQuery(),
+	getShowProductionsOfMaterialsQuery(),
 	getShowProductionsQuery()
 ];
