+++ src/neo4j/cypher-queries/company/show/show-awards-for-rights-grantor-materials.js
