+++ src/neo4j/cypher-queries/company/show/show-awards-for-sourcing-materials.js
