+++ src/neo4j/cypher-queries/company/show/show-awards-for-subsequent-version-materials.js
