+++ src/neo4j/cypher-queries/company/show/show-productions-of-materials.js
