+++ src/neo4j/cypher-queries/material/show/index.js
@@ -1,15 +1,15 @@
 import getShowQuery from './show';
 import getShowAwardsQuery from './show-awards';
-import getShowAwardsSubsequentVersionMaterialQuery from './show-awards-subsequent-version-material';
-import getShowAwardsSourcingMaterialQuery from './show-awards-sourcing-material';
+import getShowAwardsForSubsequentVersionMaterialsQuery from './show-awards-for-subsequent-version-materials';
+import getShowAwardsForSourcingMaterialsQuery from './show-awards-for-sourcing-materials';
 import getShowMaterialsQuery from './show-materials';
 import getShowProductionsQuery from './show-productions';
 
 export default () => [
 	getShowQuery(),
 	getShowAwardsQuery(),
-	getShowAwardsSubsequentVersionMaterialQuery(),
-	getShowAwardsSourcingMaterialQuery(),
+	getShowAwardsForSubsequentVersionMaterialsQuery(),
+	getShowAwardsForSourcingMaterialsQuery(),
 	getShowMaterialsQuery(),
 	getShowProductionsQuery()
 ];
