+++ src/neo4j/cypher-queries/material/show/show-awards-for-sourcing-materials.js
