+++ src/neo4j/cypher-queries/material/show/show-awards-for-subsequent-version-materials.js
