+++ src/neo4j/cypher-queries/person/show/show-awards-for-rights-grantor-materials.js
