+++ src/neo4j/cypher-queries/person/show/show-awards-for-sourcing-materials.js
