+++ src/neo4j/cypher-queries/person/show/show-productions-of-materials.js
